Collapse transport headers and functions.
