+++ benchmark/priority_queue.cpp
@@ -1,5 +1,5 @@
 #include <quicr/containers/priority_queue.h>
-#include <quicr/quic_transport.h>
+#include <quicr/transport.h>
 
 #include <benchmark/benchmark.h>
