+++ src/CMakeLists.txt
@@ -8,20 +8,20 @@ else()
 endif()
 
 target_sources (quicr PRIVATE
-    track_handler.cpp
     client.cpp
     ctrl_message_types.cpp
-    ${ctrl_messages_SOURCE}
+    fetch_track_handler.cpp
+    joining_fetch_handler.cpp
     messages.cpp
     publish_fetch_handler.cpp
     publish_track_handler.cpp
-    fetch_track_handler.cpp
-    subscribe_track_handler.cpp
     server.cpp
-    quic_transport.cpp
-    transport.cpp
+    subscribe_track_handler.cpp
+    track_handler.cpp
     transport_picoquic.cpp
-    joining_fetch_handler.cpp
+    transport.cpp
+
+    ${ctrl_messages_SOURCE}
 )
 
 set_source_files_properties(${ctrl_messages_SOURCE} PROPERTIES GENERATED TRUE)
